Painful merge... don't use multiple branches which all touch the whole project
at the same time.

Conflicts:
	Applications/MPExtended.Applications.ServiceConfigurator/Pages/SectionPluginSettings.xaml.cs
	Applications/MPExtended.Applications.ServiceConfigurator/Pages/TabAuthentication.xaml.cs
	Applications/MPExtended.Applications.ServiceConfigurator/Pages/TabBarcode.xaml.cs
	Config/Installed/Streaming.xml
	Config/MediaAccess.xml
	Config/Services.xml
	Config/Streaming.xml
	Libraries/MPExtended.Libraries.Client/MPExtended.Libraries.Client.csproj
	Libraries/MPExtended.Libraries.Service/Configuration.cs
	Libraries/MPExtended.Libraries.Service/ConfigurationContracts/Services.cs
	Libraries/MPExtended.Libraries.Service/ConfigurationContracts/Streaming.cs
	Libraries/MPExtended.Libraries.Service/MPExtended.Libraries.Service.csproj
	MPExtended.sln
	ServiceHosts/MPExtended.ServiceHosts.WebMediaPortal/MPExtended.ServiceHosts.WebMediaPortal.csproj
	Services/MPExtended.Services.StreamingService/Transcoders/FFMpeg.cs
	Services/MPExtended.Services.StreamingService/Transcoders/VLC.cs
	Services/MPExtended.Services.StreamingService/Transcoders/VLCBaseTranscoder.cs
	Services/MPExtended.Services.StreamingService/Transcoders/VLCWrapperRTSP.cs
	Services/MPExtended.Services.TVAccessService/MPExtended.Services.TVAccessService.csproj

+++ Services/MPExtended.Services.StreamingService.Interfaces/WebTranscodingInfo.cs
@@ -10,27 +10,27 @@ namespace MPExtended.Services.StreamingService.Interfaces
         /// <summary>
         /// The amount of video that has already been transcoded (0 is the start of playback position).
         /// </summary>
-        public int TranscodedTime { get; set; }
+        public long TranscodedTime { get; set; }
 
         /// <summary>
         /// The number of frames already transcoded.
         /// </summary>
-        public int TranscodedFrames { get; set; }
+        public long TranscodedFrames { get; set; }
 
         /// <summary>
         /// The position in the file at which the transcoder currently is (0 is the start of the file).
         /// </summary>
-        public int TranscodingPosition { get; set; }
+        public long TranscodingPosition { get; set; }
 
         /// <summary>
         /// The framerate at which the transcoder is currently transcoding in frames per second. Doesn't have to be the framerate of the output stream.
         /// </summary>
-        public int TranscodingFPS { get; set; } 
+        public long TranscodingFPS { get; set; } 
 
         /// <summary>
         /// The current bitrate of the output stream in kbit/s.
         /// </summary>
-        public int OutputBitrate { get; set; }
+        public long OutputBitrate { get; set; }
 
         // whether or not getting the transcoding info is supported
         public bool Supported { get; set; }
